Source/requirementsYOLOP.txt
absl-py==0.13.0
appdirs==1.4.3
argon2-cffi==20.1.0
astroid==2.4.0
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==19.10b0
bleach==3.3.1
bumpversion==0.5.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
config==0.5.0.post0
coverage==5.1
cryptography==3.4.7
cv-utils==0.2.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.2
docutils==0.17.1
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.10.0
identify==2.2.11
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.3
importlib-resources==5.2.2
imutils==0.5.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplot==0.1.9
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nodeenv==1.6.0
notebook==6.4.0
numpy==1.18.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.3
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkginfo==1.7.1
platformdirs==2.2.0
pluggy==0.13.1
pre-commit==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pylint==2.5.0
pyloco==0.0.139
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==5.4.1
pytest-cov==2.8.1
pytest-mock==3.1.0
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
readme-renderer==29.0
regex==2020.4.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.4.1
SecretStorage==3.3.1
Send2Trash==1.7.1
SimpleWebSocketServer==0.1.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-addons==0.10.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2020.9.3
toml==0.10.0
tornado==6.1
tqdm==4.62.0
traitlets==4.3.3
twine==3.4.2
typed-ast==1.4.1
typeguard==2.12.1
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.6
ushlex==0.99.1
virtualenv==20.7.0
visdom==0.1.8.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
